Rename median_vareps.py to median.py
